+++ packages/lesswrong/components/common/AnalyticsClient.tsx
@@ -3,7 +3,7 @@ import { registerComponent } from '../../lib/vulcan-lib/components';
 import { clientContextVars, flushClientEvents } from '../../lib/analyticsEvents';
 import { useCurrentUser } from './withUser';
 import withErrorBoundary from './withErrorBoundary';
-import { ABTestGroupsUsedContext } from '../../lib/abTestImpl';
+import { ABTestGroupsUsedContext } from '@/components/common/sharedContexts';
 import { CLIENT_ID_COOKIE } from '../../lib/cookies/cookies';
 import { useCookiesWithConsent } from '../hooks/useCookiesWithConsent';
 import { isLWorAF } from '../../lib/instanceSettings';
